TP01

+++ Tp01/calculadora.h
@@ -6,4 +6,4 @@ int suma(float numeroUno, float numeroDos);
 int resta(float numeroUno, float numeroDos);
 int multiplicacion(float numeroUno, float numeroDos);
 int division(float numeroUno, float numeroDos);
-int factorial(int numeroUno);
+int factorial(float numeroUno);
